Rename the class and move the file.
