+++ MetaphysicsIndustries.Solus.Test/ExpressionsT/ExpressionCheckerT/FunctionsT/DeriveOperatorT/CallTest.cs
@@ -1,4 +1,3 @@
-
 /*
  *  MetaphysicsIndustries.Solus
  *  Copyright (C) 2006-2022 Metaphysics Industries, Inc., Richard Sartor
@@ -20,14 +19,12 @@
  *
  */
 
-using MetaphysicsIndustries.Solus.Evaluators;
 using MetaphysicsIndustries.Solus.Expressions;
 using MetaphysicsIndustries.Solus.Functions;
-using MetaphysicsIndustries.Solus.Macros;
 using NUnit.Framework;
 
 namespace MetaphysicsIndustries.Solus.Test.ExpressionsT.ExpressionCheckerT.
-    MacrosT.DeriveMacroT
+    FunctionsT.DeriveOperatorT
 {
     [TestFixture]
     public class CallTest
@@ -38,7 +35,7 @@ namespace MetaphysicsIndustries.Solus.Test.ExpressionsT.ExpressionCheckerT.
             // given
             var parser = new SolusParser();
             var expr = new FunctionCall(
-                new Literal(DeriveMacro.Value),
+                new Literal(DeriveOperator.Value),
                 parser.GetExpression("3*x^2+5*x+7"),
                 new VariableAccess("x"));
             var env = new SolusEnvironment();
